% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pgscatalog_ancestry_info}
\alias{pgscatalog_ancestry_info}
\title{Get ancestry metadata}
\usage{
pgscatalog_ancestry_info()
}
\value{
A data.frame with columns: acronym, ancestry, pop_group, color.
}
\description{
Returns a data.frame with acronyms, full ancestry names, population group
abbreviations, and hex colors.
}
\examples{
pgscatalog_ancestry_info()
}
