% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pgscatalog_palettes}
\alias{pgscatalog_palettes}
\title{List available PGS Catalog palettes}
\usage{
pgscatalog_palettes()
}
\value{
A named list of available palettes.
}
\description{
List available PGS Catalog palettes
}
\examples{
pgscatalog_palettes()
}
